+++ src/components/calcite-example/calcite-example.e2e.ts
@@ -1,4 +1,5 @@
 import { newE2EPage } from "@stencil/core/testing";
+import { HYDRATED_ATTR } from "../../tests/commonTests";
 
 describe.skip("calcite-example", () => {
   it("renders", async () => {
@@ -6,6 +7,6 @@ describe.skip("calcite-example", () => {
 
     await page.setContent("<calcite-example></calcite-example>");
     const element = await page.find("calcite-example");
-    expect(element).toHaveAttribute("calcite-hydrated");
+    expect(element).toHaveAttribute(HYDRATED_ATTR);
   });
 });