+++ src/components/calcite-tab-title/calcite-tab-title.e2e.ts
@@ -1,4 +1,5 @@
 import { newE2EPage } from "@stencil/core/testing";
+import { HYDRATED_ATTR } from "../../tests/commonTests";
 
 describe("calcite-tab-title", () => {
   it("renders", async () => {
@@ -6,6 +7,6 @@ describe("calcite-tab-title", () => {
 
     await page.setContent("<calcite-tab-title></calcite-tab-title>");
     const element = await page.find("calcite-tab-title");
-    expect(element).toHaveAttribute("calcite-hydrated");
+    expect(element).toHaveAttribute(HYDRATED_ATTR);
   });
 });